New in-memory store
Completely rework the in-memory store and change tracking. Implement an MVCC to handle concurrent updates to a single row. No longer rely on classes from the System.Data namespace.
All write operations now create a transaction to act as a safeguard in case of an error, as well as to support the MVCC.

Many bug fixes as well as boilerplate reduction


* Add versioning for rows

* Add tracking for row changes

* Use change tracking everywhere

* Rework the store mechanism so that it relies on buffers rather than System.Data

* Finish tables

* Add implicit transactions

* Rename a method and invoke a validator

* Make an internal transaction class

* Create interfaces that allow other assemblies to access table information

* Fix compile errors in the main project

Binding for nullable value types is broken... as well as the tests

* Fix the nullability issue, and many other bugs

Tests are still broken though

* Remove unused resources

* Fix transactions

* Wee refactor

* Fix transactions and another wee refactor

* Extract a method

* Fix tests for the sqlite provider

* Change code style prefs

* Fix project name and exclude it from code coverage

* Fix project name... again and extract common test code into a separate project

* Move the class to the appropriate namespace

* Various test and source fixes

Honestly, I have no idea what I did here, so... yeah.

* Exclude test projects from code coverage